requirements.txt
pandas
openpyxl
plotext
networkx
pyvis
yfinance
fpdf2
matplotlib
streamlit
plotly
